feat: comprehensive updates across client, channels, packaging, and build system
- Android client: major refactor of freeRDPCore module, new keyboard accessibility service,
  print job monitoring, clipboard image provider, gesture detection improvements
- Channels: updates to audin, cliprdr, printer (Android support), rdpgfx, rdpsnd
- SDL client: touch handling improvements for SDL2 and SDL3
- X11/Windows clients: clipboard, keyboard, GDI updates
- Core libfreerdp: codec improvements (H.264, SSE), connection handling, orders, surface,
  gateway, fastpath, gfx utilities
- WinPR: crypto, file, input, interlocked, path, pipe, SSPI updates with new tests
- Packaging: distro package container builds (Dockerfile, scripts for Arch/Debian/RPM),
  Flatpak, RPM spec, debian rules updates
- CI/CD: GitHub workflow updates for FreeBSD, timezone
- CMake: new filesystem detection, config options, Android/OpenH264 patches
- Proxy server: channel modules, bitmap filter, context, input, update handling
- Shadow server: client improvements

+++ libfreerdp/codec/dsp.c
@@ -320,7 +320,7 @@ static const INT16 ima_step_size_table[] = {
 	12635, 13899, 15289, 16818, 18500, 20350, 22385, 24623, 27086, 29794, 32767
 };
 
-static inline void dsp_ima_clamp_step(ADPCM* WINPR_RESTRICT adpcm, unsigned int channel)
+static inline void dsp_ima_clamp_step(ADPCM* WINPR_RESTRICT adpcm, size_t channel)
 {
 	WINPR_ASSERT(adpcm);
 	if (adpcm->ima.last_step[channel] < 0)
@@ -413,7 +413,7 @@ static BOOL freerdp_dsp_decode_ima_adpcm(FREERDP_DSP_CONTEXT* WINPR_RESTRICT con
 			    (INT16)(((UINT16)(*src)) | (((UINT16)(*(src + 1))) << 8));
 			context->adpcm.ima.last_step[0] = (INT16)(*(src + 2));
 
-			dsp_ima_clamp_step(&context->adpcm, 0);
+			dsp_ima_clamp_step(&context->adpcm, 0u);
 
 			src += 4;
 			size -= 4;
@@ -427,7 +427,7 @@ static BOOL freerdp_dsp_decode_ima_adpcm(FREERDP_DSP_CONTEXT* WINPR_RESTRICT con
 				    (INT16)(((UINT16)(*src)) | (((UINT16)(*(src + 1))) << 8));
 				context->adpcm.ima.last_step[1] = (INT16)(*(src + 2));
 
-				dsp_ima_clamp_step(&context->adpcm, 1);
+				dsp_ima_clamp_step(&context->adpcm, 1u);
 				src += 4;
 				size -= 4;
 				out_size -= 16;
